Organize into multiple files
- Allow LTILaunchValidator to be used without needing to use all
  the JupyterHub related things!
- Use a specific LTLaunchValidationError than just web.HTTPError
